refactor: CarNamesValidator와 TryCountValidator를 Validator로 통합
